add tx id into proof message - has some problems however... need to make it work async
